third_party/WebKit/LayoutTests/fast/xpath/xpath-detached-import-assert-expected.txt

CONSOLE WARNING: The behavior of importNode() with no boolean argument is about to change from doing a deep clone to doing a shallow clone.  Make sure to pass an explicit boolean argument to keep your current behavior.
A test case from https://bugs.webkit.org/show_bug.cgi?id=84793.

Test passes if it doesn't hit ASSERT(parents.size() >= depth + 1).

PASS result.numberValue is NaN
PASS successfullyParsed is true

TEST COMPLETE
